teste 2

+++ features/steps/soma_steps.py
@@ -3,8 +3,8 @@ from soma import somar
 
 @given('que eu tenho dois números: {num1} e {num2}')
 def definir_numeros(context, num1, num2):
-    context.num1 = int(num1)
-    context.num2 = int(num2)
+    context.num1 = float(num1)  # Alterado para float
+    context.num2 = float(num2)  # Alterado para float
 
 @when('eu os somo')
 def somar_numeros(context):
@@ -12,4 +12,4 @@ def somar_numeros(context):
 
 @then('o resultado deve ser {resultado_esperado}')
 def verificar_resultado(context, resultado_esperado):
-    assert context.resultado == int(resultado_esperado)
+    assert context.resultado == float(resultado_esperado)
